move nftcache table and related actions over to the dacdirectory contract
